refactor(cli): extract gateway runtime
